fix: use signed loop variables in OpenMP for MSVC compatibility
MSVC OpenMP 2.0 requires signed integral types for parallel for loop
variables. Changed size_t to int64_t in all OpenMP parallel for loops.

+++ cpp/trimci_core/matfree_davidson.cpp
@@ -5,6 +5,7 @@
 #include <algorithm>
 #include <cstring>
 #include <cstddef>
+#include <cstdint>
 #include <chrono>
 #include <cstdio>
 #include <cmath>
@@ -284,7 +285,7 @@ void precompute_connections(InverseMapData& data,
     #ifdef _OPENMP
     #pragma omp parallel for schedule(dynamic, 256)
     #endif
-    for (size_t d = 0; d < n_dets; ++d)
+    for (int64_t d = 0; d < (int64_t)n_dets; ++d)
         data.h_diag_det[d] = compute_diagonal(
             data.all_dets[d].alpha, data.all_dets[d].beta, h1, eri, n_orb);
 
@@ -309,7 +310,7 @@ void precompute_connections(InverseMapData& data,
 #ifdef _OPENMP
         #pragma omp for schedule(dynamic, 64)
 #endif
-        for (size_t d_idx = 0; d_idx < n_dets; ++d_idx) {
+        for (int64_t d_idx = 0; d_idx < (int64_t)n_dets; ++d_idx) {
             uint64_t alpha_d = data.all_dets[d_idx].alpha;
             uint64_t beta_d = data.all_dets[d_idx].beta;
             Det det_d = {alpha_d, beta_d};
@@ -444,7 +445,7 @@ void compute_Sv(const double* v, double* result, int n_basis,
 #ifdef _OPENMP
         #pragma omp for schedule(dynamic, 256)
 #endif
-        for (size_t d = 0; d < n_dets; ++d) {
+        for (int64_t d = 0; d < (int64_t)n_dets; ++d) {
             const ContribVec& contribs = *data.all_dets[d].contribs;
             // Gather: sigma = sum c_mu * v[mu]
             double sigma = 0.0;
@@ -500,7 +501,7 @@ void compute_Hv(const double* v, double* result, int n_basis,
 #ifdef _OPENMP
         #pragma omp for schedule(static)
 #endif
-        for (size_t d = 0; d < n_dets; ++d) {
+        for (int64_t d = 0; d < (int64_t)n_dets; ++d) {
             double val = data.h_diag_det[d] * sigma[d];
             for (const auto& [mu, c_mu] : *data.all_dets[d].contribs)
                 local[mu] += c_mu * val;
@@ -510,7 +511,7 @@ void compute_Hv(const double* v, double* result, int n_basis,
 #ifdef _OPENMP
         #pragma omp for schedule(static)
 #endif
-        for (size_t c = 0; c < n_conn; ++c) {
+        for (int64_t c = 0; c < (int64_t)n_conn; ++c) {
             const auto& conn = data.connections[c];
             double h = conn.h_elem;
             double s1 = sigma[conn.d1], s2 = sigma[conn.d2];
@@ -562,7 +563,7 @@ void compute_diagonals(Eigen::VectorXd& H_diag, Eigen::VectorXd& S_diag,
 #ifdef _OPENMP
         #pragma omp for schedule(dynamic, 256)
 #endif
-        for (size_t d = 0; d < n_dets; ++d) {
+        for (int64_t d = 0; d < (int64_t)n_dets; ++d) {
             const ContribVec& contribs = *data.all_dets[d].contribs;
             double h_dd = data.h_diag_det[d];
             for (const auto& [mu, c_mu] : contribs) {
